Moved win2d stuff to Win2DHandler class
