--- Classes/Win2DHandeler.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.Graphics.Canvas;
-using Microsoft.Graphics.Canvas.UI.Xaml;
-
-namespace LudoGame.Classes
-{
-    class Win2DHandeler
-    {
-        private CanvasAnimatedControl canvas;
-        public void Update()
-        {
-
-        }
-        public void Draw() 
-        { 
-        
-        }
-    }
-}
